task_manager/requirements.txt
pytest==7.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1